docs(lib-dynamodb): fix import path for api extractor compatibility
